+++ common/src/main/java/com/lowdragmc/photon/command/BlockEffectCommand.java
@@ -123,7 +123,7 @@ public class BlockEffectCommand extends EffectCommand {
             if (fx != null) {
                 var effect = new BlockEffect(fx, handler.getLevel(), pos);
                 effect.setOffset(offset.x, offset.y, offset.z);
-                effect.setRotation(rotation.x, offset.y, offset.z);
+                effect.setRotation(rotation.x, rotation.y, rotation.z);
                 effect.setScale(scale.x, scale.y, scale.z);
                 effect.setDelay(delay);
                 effect.setForcedDeath(forcedDeath);