+++ CHANGELOG.md
@@ -1,4 +1,3 @@
 ## ChangeLog
 
-* Fixed fx exported by the latest photon could not be displayed.
-* Fixed compatibility with previous entity fx command.
+* Fixed incorrect rotation